Refactoring lexer
